event-gateway/requirements.txt
# Dependências para o serviço de gateway de eventos
flask
paho-mqtt
gunicorn